Restructure project:
- Convert to Gradle
- Add a minimal CircleCI configuration
- Add a test-host to able to use AndroidJUnitRunner
- Make the demo/example a subproject (so it get recompile upon SDK build)
- Temporarily disable `sendToTwoTableWithLimitedUploadEvents()` (failed constantly)
