--- example/settings.gradle
@@ -1,1 +1,0 @@
-include ':td-android-sdk-demo'
