+++ example/src/main/java/com/treasuredata/android/demo/InstallationReceiver.java
