+++ example/src/main/java/com/treasuredata/android/demo/MainActivity.java
