--- example/td-android-sdk-demo/build.gradle
@@ -1,30 +1,0 @@
-apply plugin: 'com.android.application'
-
-android {
-    compileSdkVersion 28
-    buildToolsVersion "28.0.3"
-
-    defaultConfig {
-        applicationId "com.treasuredata.android.demo"
-        minSdkVersion 14
-        targetSdkVersion 28
-        versionCode 1
-        versionName "1.0"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.txt'
-        }
-    }
-}
-
-repositories {
-    jcenter()
-    mavenCentral()
-}
-
-dependencies {
-    api 'com.treasuredata:td-android-sdk:0.1.19'
-}
