+++ settings.gradle
@@ -1,0 +1,3 @@
+rootProject.name = 'td-android-sdk'
+
+include ':example', ':test-host'
