+++ test-host/README.md
@@ -1,0 +1,3 @@
+To be used a host runtime for unit tests that involve platform API.
+
+For the demo application, see `../example`
